Move step into its own module
