+++ gwcs/wcs/__init__.py
@@ -1,5 +1,6 @@
 from ._exception import GwcsBoundingBoxWarning, NoConvergence
-from ._wcs import WCS, Step
+from ._step import Step
+from ._wcs import WCS
 
 __all__ = [
     "WCS",
